moved effects back into client context, as it now works!
