+++ src/components/Client/context.jsx
@@ -1,5 +1,9 @@
-import React, { createContext, useReducer } from "react";
-import { reducer } from "./";
+import React, { createContext, useReducer, useEffect } from "react";
+import { 
+	ClientContext as Client,
+	ACTIONS, 
+	reducer 
+} from "./";
 import UTILS from "SHARED/utils.js";
 
 const initialState = {
@@ -24,6 +28,30 @@ function ClientProvider(props){
 	const { children } = props;
 
 	const [state, dispatch] = useReducer(reducer, initialState);
+
+	function updateResizeListeners(){
+		window.addEventListener("resize", syncViewSize);
+		return () => {
+			window.removeEventListener("resize", syncView)
+		};
+	}// updateResizeListeners
+
+
+	function syncViewSize(){
+		dispatch({
+			type: ACTIONS.UPDATE_VIEW_SIZE
+		});
+	}// syncViewSize
+
+
+	// EFFECTS
+	// ------------------------
+
+	//set device dimensions on mount
+	useEffect(syncViewSize, []); 
+
+	//add/remove resize listeners
+	useEffect(updateResizeListeners, []);
 
 	return (
 		<ClientContext.Provider value={ { state, dispatch } }>
